+++ java-8-test/src/test/java/fi/vincit/multiusertest/ChainedTest.java
@@ -12,12 +12,10 @@ import fi.vincit.multiusertest.annotation.TestUsers;
 import fi.vincit.multiusertest.configuration.ConfiguredTest;
 import fi.vincit.multiusertest.rule.expection.value.ExpectValueOf;
 import fi.vincit.multiusertest.runner.junit.MultiUserTestRunner;
-import fi.vincit.multiusertest.runner.junit.framework.BlockMultiUserTestClassRunner;
 import fi.vincit.multiusertest.util.LoginRole;
 
 @TestUsers(creators = {"role:ROLE_SUPER_ADMIN", "role:ROLE_ADMIN", "role:ROLE_USER", },
-        users = {"role:ROLE_ADMIN", "role:ROLE_USER", "role:ROLE_VISITOR"},
-        runner = BlockMultiUserTestClassRunner.class)
+        users = {"role:ROLE_ADMIN", "role:ROLE_USER", "role:ROLE_VISITOR"})
 @RunWith(MultiUserTestRunner.class)
 public class ChainedTest extends ConfiguredTest {