+++ java-8-test/src/test/java/fi/vincit/multiusertest/Java8SmokeTest.java
@@ -12,11 +12,9 @@ import org.junit.runner.RunWith;
 import fi.vincit.multiusertest.annotation.TestUsers;
 import fi.vincit.multiusertest.configuration.ConfiguredTest;
 import fi.vincit.multiusertest.runner.junit.MultiUserTestRunner;
-import fi.vincit.multiusertest.runner.junit.framework.BlockMultiUserTestClassRunner;
 import fi.vincit.multiusertest.util.LoginRole;
 
-@TestUsers(creators = {"role:ROLE_ADMIN"}, users = "role:ROLE_USER",
-        runner = BlockMultiUserTestClassRunner.class)
+@TestUsers(creators = {"role:ROLE_ADMIN"}, users = "role:ROLE_USER")
 @RunWith(MultiUserTestRunner.class)
 public class Java8SmokeTest extends ConfiguredTest {